requirements.txt
flask
nltk
python-dotenv
requests
trafilatura
WTForms
